refactor(sdk): moved `HMSManagerModule` from "classes" dir to new "modules" dir
