+++ packages/react-native-hms/src/modules/HMSManagerModule.ts
